// src/appointments/enums/openai.ts
export enum OpenAIModel {
  GPT_3_5_TURBO_0125 = "gpt-3.5-turbo-0125",
  WHISPER_1 = "whisper-1",
}

export enum OpenAIModelType {
  JSON_OBJECT = "json_object",
}

export enum Socket {
  UPGRADE = "upgrade",
  CONNECTION = "connection",
  ERROR = "error",
  CLOSE = "close",
  MESSAGE = "message",
}